third_party/blink/web_tests/fast/lists/ol-nested-list-dynamic-remove-expected.txt

This tests that list item markers do not change when removing existing items from a nested list.

1 One
2 Two
3 Three
4 Four
5 Five
1 One
2 Two
